+++ lib/repositories/db/enums/customer_key_type.dart
@@ -1,0 +1,10 @@
+﻿enum CustomerKeyType {
+  externiId('ext'),
+  ic('ic'),
+  nazev('name');
+
+  const CustomerKeyType(this.prefix);
+  final String prefix;
+
+  String format(String value) => '$prefix:$value';
+}