+++ lib/repositories/db/enums/db_table.dart
@@ -1,0 +1,10 @@
+﻿enum DbTable {
+  zakaznici('zakaznici'),
+  operace('operace'),
+  materialy('materialy'),
+  profily('profily'),
+  customerProfiles('customer_profiles');
+
+  const DbTable(this.tableName);
+  final String tableName;
+}