+++ lib/repositories/db_repository.dart
@@ -1,0 +1,5 @@
+﻿import 'package:sqflite_common_ffi/sqflite_ffi.dart';
+
+abstract class DbRepository {
+  Future<Database> get database;
+}